ajuste de float par int

+++ lista05/lista05_lista01_questao42.py
@@ -19,7 +19,7 @@ sum26_50 =0
 sum51_75 = 0
 sum76_100 = 0
 
-number = float(input("Digite qualquer número inteiro positivo (para sair digite um númeronegativo)"))
+number = int(input("Digite qualquer número inteiro positivo (para sair digite um númeronegativo)"))
 
 while number > 0:
     if 0 < number <= 25:
